Update formula to include distnoted patch
Emacs 24.3 is known to cause issues on OSX Mavericks, especially when
waking from sleep, which manifest as the distnoted process hogging CPU
cycles. There's a patch in trunk (24.4), which has been backported in
the upstream homebrew formula.

I've merged in all the upstream homebrew formula changes, which mainly
reflect changes in core homebrew itself. The only significant change is
the addition of this patch (referred as a 'memory leak' patch).